- MainBanner, Filter, ChatList 컴포넌트화 - Main.js 컴포넌트화된거 적용
