+++ src/components/mainBanner/MainBanner.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+
+
+export default class MainBanner extends Component {
+  render() {
+    return (
+      <section className="main__banner">
+        <article>
+          <div>
+            <h2>Beagle</h2>
+            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed ...</p>
+          </div>
+        </article>
+      </section>
+    )
+  }
+}
